// src/interfaces/InterfaceAsistenciaMatriculados.tsx
export interface InterfaceAsistenciaMatriculados {
    apellidos:            string;
    idmateriasporpersona: number;
    asistencia:           boolean;
    etiqueta_reconocer:   string;
    numero_telefono:      string;
    idpersona:            number;
    correo:               string;
    genero:               string;
    idmatriculacion:      number;
    nombres:              string;
}